Add statistics package
Statistics package who will provide data statistics on:
1. PhtNodes, their usage, their number of keys
2. Nodes, their usage
3. Operations requested by the client
4. Operations requested by PhtNodes (especially for
range queries and lookups)

+++ pht/src/peersim/pht/messages/PMRangeQuery.java
@@ -14,7 +14,7 @@ import java.util.List;
  * for a parallel lookup) and send a general acquittal to the initiator node.
  */
 public class PMRangeQuery {
-    private int count;
+    private long count;
     private boolean end;
     private final String keyMin;
     private final String keyMax;
@@ -24,7 +24,7 @@ public class PMRangeQuery {
         this(keyMin, keyMax, 0);
     }
 
-    public PMRangeQuery(String keyMin, String keyMax, int count) {
+    public PMRangeQuery(String keyMin, String keyMax, long count) {
         this.keyMin = keyMin;
         this.keyMax = keyMax;
         this.count  = count;
@@ -71,7 +71,7 @@ public class PMRangeQuery {
         return keyMax;
     }
 
-    public int getCount() {
+    public long getCount() {
         return this.count;
     }
